'fixed' the plugin but lost some functionality
See the diff for the exact line number but in category_meta.py
there is a line in the `make_category` function:

category.slug = slug

I'm not 100% certain of the author of category_meta's intentions
but I believe this line attempts to change the category slug to the
name of the directory the post exists in.

By removing this line it breaks that functionality of the plugin but
everything else works as expected.

For example you can see that the category description in
`category/category.html` now appears as expected.